Use strings for checkbox plugin settings
Redmine doesn't use booleans for checkboxes in plugin settings.
Instead Redmine either stores text value if checkbox is checked and
doesn't store the value if checkbox isn't checked. Therefore in order
to check for the value properly we should compare with the checkbox
input value for true and pass nil when