Pass author to wiki update job
Redmine may change author in the page.content to something different
in some cases. This change is not persisted and if we don't pass it
to the sidekiq worker additionally it will be lost and the page update
will be rendered with the author value from the database. Since it
differs from the intended value in the cases when Redmine replaces it
this may result in wrong values in rendering. One of such cases is
when the user changes the parent page of a wiki page to something
different without changing page content. In this case page.content.author
points to the author of the last content change but Redmine changes it
to the user who performed the parent page change without creating
new content intance and without persisting the author change to
the last content object.